backend/requirements.txt
Flask
requests
google-auth-oauthlib
psycopg2-binary
python-dotenv
gunicorn
Flask-Cors
Pillow
google-generativeai
gspread
cryptography
PyJWT